backend/requirements.txt
# svdeeq-backend/requirements.txt

fastapi==0.111.0
uvicorn[standard]==0.29.0
pydantic==2.7.1
pydantic-settings==2.2.1

# Supabase
supabase==2.4.6

# Google Gemini
google-generativeai==0.5.4

# HTTP client (for Evolution API calls)
httpx==0.27.0

# Background tasks queue
anyio==4.3.0

# Env
python-dotenv==1.0.1
